+++ crates/environ/src/compile/module_environ.rs
